fixing the bug introduced in [1369]
don't allow someone to arbitrarily delete files when they don't have access to

+++ src/net/sf/drftpd/master/command/plugins/DataConnectionHandler.java
@@ -1293,7 +1293,9 @@ public class DataConnectionHandler implements CommandHandler, CommandHandlerFact
                 //_transfer is already set up by doPASV()
             } else {
             	// reset(); already done in finally block
-            	_transferFile.delete();
+            	if (isStor) {
+            		_transferFile.delete();
+            	}
                 return Reply.RESPONSE_503_BAD_SEQUENCE_OF_COMMANDS;
             }
